+++ packages/apollo-client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### vNext
 - Support conditional refetches for mutations
+- Ensure network errors don't cause unhandled rejections on cache-and-network policies
 
 ### 2.0.0-beta.5
 - Fix argument for FragmentMatcher